Version Log + account block
Додано:
- журнал змін
- динамічна зміна версії
- сайт буде запитувати логін при виборі "Історія оновлень"
- Запит на вихід із кабінету

+++ js/account.js
@@ -24,7 +24,9 @@ document.addEventListener('DOMContentLoaded', function () {
 });
 
 function Exit() {
-    sessionStorage.clear();
-    location.reload();
+    if (confirm("Do you confirm log out?") == true) {
+        sessionStorage.clear();
+        location.reload();
+    } 
 
 }
